+++ src/GameSvr/Player/PlayObject.Operate.cs
@@ -736,7 +736,7 @@ namespace GameSvr.Player
                 {
                     if (baseObject.Death && !baseObject.Skeleton && baseObject.Animal)
                     {
-                        var n10 = M2Share.RandomNumber.Random(16) + 5;
+                        var n10 = (byte)(M2Share.RandomNumber.Random(16) + 5);
                         var n14 = (ushort)(M2Share.RandomNumber.Random(201) + 100);
                         baseObject.BodyLeathery -= n10;
                         baseObject.MeatQuality -= n14;
